refactor(cli): prepare for mobile implementation
